+++ src/icons.js
@@ -9,6 +9,10 @@ const Icons = (() => {
   const NS = 'http://www.w3.org/2000/svg';
 
   const PATHS = {
+    'archive':
+      '<rect width="20" height="5" x="2" y="3" rx="1"/> <path d="M4 8v11a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8"/> <path d="M10 12h4"/>',
+    'archive-restore':
+      '<rect width="20" height="5" x="2" y="3" rx="1"/> <path d="M4 8v11a2 2 0 0 0 2 2h2"/> <path d="M20 8v11a2 2 0 0 1-2 2h-2"/> <path d="m9 15 3-3 3 3"/> <path d="M12 12v9"/>',
     'check':
       '<path d="M20 6 9 17l-5-5"/>',
     'chevron-down':